big ol cleanup
