+++ graphix/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 graphix_macros = { path = "../graphix_macros" }
+graphix_models = { path = "../graphix_models" }
 hcl-rs = { version = "0.16.9", features = ["perf"] }
 serde = "1.0.201"
 serde_json = "1.0.117"
