--- graphix/src/entity/attribute.rs
@@ -1,8 +1,0 @@
-use strum::Display;
-
-#[derive(Debug, Display, Copy, Clone)]
-pub enum Attribute {
-    Immutable,
-    Unique,
-    Nullable,
-}
