+++ graphix_macros/Cargo.toml
@@ -8,6 +8,7 @@ proc-macro = true
 name = "graphix_macros"
 
 [dependencies]
+graphix_models = { path = "../graphix_models" }
 proc-macro2 = "1.0.82"
 quote = "1.0.36"
 strum = { version = "0.26.2", features = ["derive"] }
