+++ graphix_models/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "graphix_models"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+quote = "1.0.36"
+syn = "2.0.62"
+proc-macro2 = "1.0.82"
+hcl-rs = { version = "0.16.9", features = ["perf"] }
