+++ graphix_models/src/entity.rs
@@ -1,20 +1,9 @@
+use crate::descriptor::{EntityDescriptor, EntityFieldDescriptor};
 use hcl::Block;
-
-mod attribute;
-pub mod field;
-mod field_type;
-
-#[derive(Debug)]
-pub struct EntityDescriptor {
-    pub name: String,
-    pub table_name: String,
-    pub schema_name: String,
-    pub fields: Vec<field::EntityFieldDescriptor>,
-}
 
 pub trait Entity {
     fn entity_descriptor(&self) -> EntityDescriptor;
-    fn as_atlas_hcl(&self) -> hcl::Block {
+    fn as_atlas_hcl(&self) -> Block {
         let desc = self.entity_descriptor();
 
         let mut builder = Block::builder("table")
