+++ graphix_models/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod descriptor;
+pub mod sql;
+pub mod entity;
